rename interpolation_join module
